feat: moved the logic from FRCamera to FaceRecognition
