+++ screens/FaceRecognition.js
@@ -1,12 +1,19 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 import { Camera } from 'expo-camera';
 import { useSelector } from 'react-redux';
+
+import base64ToArrayBuffer from 'base64-arraybuffer'; // for converting base64 images to array buffer
+import * as AzureAPI from './../services/azureAPI';
+
 import TextMessage from '../components/FaceRecognition/TextMessage';
 import FRCamera from './../components/FaceRecognition/FRCamera';
 import FaceSquares from './../components/FaceRecognition/FaceSquares';
 
 function FaceRecognition() {
+  // CONSTANTS
+  const SHOW_QUOTE_TIME = 5000; // [ms]
+
   // LOCAL STATE
   const [eFaceRecState] = useState({
     TAKE_SELFIE: 1,
@@ -21,6 +28,9 @@ function FaceRecognition() {
   // GLOBAL STATE
   const selectedDoor = useSelector((state) => state.selectedDoor);
 
+  // CAMERA REF
+  const cam = useRef();
+
   // HOOKS
   useEffect(() => {
     (async () => {
@@ -28,6 +38,56 @@ function FaceRecognition() {
       setHasPermission(status === 'granted');
     })();
   }, []);
+
+  // HELPER FUNCTIONS
+  /**
+   * Callback called when a face is detected, it updates the local state.
+   *
+   * @param {array} faces, array containing all the faces detected by the camera.
+   */
+  const _handleFacesDetected = ({ faces }) => {
+    setDetectedFaces(faces);
+  };
+
+  /**
+   * Callback called when a picture is taken.
+   */
+  const _takePicture = async () => {
+    setFaceRecState(eFaceRecState.CHECKING_FACE);
+
+    const option = {
+      quality: 0.25,
+      base64: true,
+    };
+    try {
+      const picture = await cam.current.takePictureAsync(option);
+      _checkPicture(picture);
+    } catch (error) {
+      console.log('error :>> ', error);
+    }
+  };
+
+  /**
+   * Check if a picture contains a face calling the Azure API.
+   *
+   * @param {object} picture, picture taken with the camera.
+   */
+  const _checkPicture = async (picture) => {
+    const octetStream = base64ToArrayBuffer.decode(picture.base64);
+    const faceDetectRes = await AzureAPI.detectFace(octetStream);
+
+    // TODO: send the face id and the selected door to the back end
+    // Wait for the answer
+    // If the user is allowed send the request to open the door
+    if (faceDetectRes.length === 0) {
+      setFaceRecState(eFaceRecState.FACE_NOT_DETECTED);
+    } else {
+      setFaceRecState(eFaceRecState.FACE_DETECTED);
+    }
+    setTimeout(() => {
+      setFaceRecState(eFaceRecState.TAKE_SELFIE);
+    }, SHOW_QUOTE_TIME);
+  };
 
   // RENDER
   if (hasPermission === null) {
@@ -38,10 +98,10 @@ function FaceRecognition() {
     return (
       <View style={styles.container}>
         <FRCamera
-          eFaceRecState={eFaceRecState}
-          setFaceRecState={setFaceRecState}
           detectedFaces={detectedFaces}
-          setDetectedFaces={setDetectedFaces}
+          _handleFacesDetected={_handleFacesDetected}
+          _takePicture={_takePicture}
+          cam={cam}
         />
         <FaceSquares detectedFaces={detectedFaces} />
         <TextMessage
